pyproject.toml
[project]
name = "anipy"
version = "0.4.0"
description = "animeshka :P"
requires-python = ">=3.12"
dependencies = [
    "Pycryptodome",
    "aiohttp",
    "requests",
    "flask[async]>=3.1.1",
    "flask-cors>=6.0.1",
    "parsel>=1.11.0",
]

[project.scripts]
anipy = "anipy.cli.main:main"
anipyweb = "anipy.web.app:main"
